ufficiourbanistica/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
design-django-theme==1.4.5
Django==3.2.6
django-appconf==1.0.4
django-compressor==2.4.1
django-cors-headers==3.8.0
django-environ==0.4.5
django-filebrowser==3.14.1
django-grappelli==2.15.1
django-sass-processor==1.0.1
django-tinymce==3.3.0
libsass==0.21.0
lxml==4.6.3
Pillow==8.3.1
pkg_resources==0.0.0
pytz==2021.1
rcssmin==1.0.6
rjsmin==1.1.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
